:recycle: feat: youtubeHelpers file
Move youtubeClient to youtubeHelpers file, change files affected by refactor
