+++ src/pages/Home/Home.page.test.jsx
@@ -12,7 +12,7 @@ import { MemoryRouter } from 'react-router';
 
 import HomePage from './index';
 import VideosProvider from '../../providers/VideoProvider';
-import { youtubeClient } from '../../utils/helpers';
+import { youtubeClient } from '../../utils/helpers/youtubeHelpers';
 import videos from '../../mocks/videos.json';
 
 const build = () => {
