+++ src/utils/hooks/useYoutubeApi.js
@@ -1,7 +1,8 @@
 import { useContext, useState } from 'react';
 import VideosContext from '../../providers/VideoProvider/VideoContext';
 import { setVideosAction } from '../../providers/VideoProvider/VideoProvider.actions';
-import { youtubeClient, buildQueryParams } from '../helpers';
+import { buildQueryParams } from '../helpers/helpers';
+import { youtubeClient } from '../helpers/youtubeHelpers';
 
 const useYoutubeApi = () => {
   const [data, setData] = useState(null);
